[MWB] - moving Common.MachineStuff.cs to MachineStuff.cs - #35155
